requirements.txt
requests====2.31.0
joblib==1.3.2
googlesearch-python==1.2.3
pydantic==2.6.0
